+++ modules/documentation/docs/changelog.md
@@ -4,7 +4,9 @@
 
 
 - \[ops\] Add persistent storage for Grafana
-
+- \[contracts\] Add `transactionHash` to withdrawal commitment
+- \[store\] Add `transferDefinition` to filter options on `getTransfers`
+- \[store\] Add `getWithdrawalCommitmentByTransactionHash` method
 
 ## 0.1.15-beta.12